requirements.txt
requests==2.28.1
pytest==7.1.3
isort==5.10.1
selenium==3.141.0
webdriver-manager==3.8.3
Pillow~=9.2.0


